+++ src/rogallo/widgets/command_line/widget.py
@@ -41,11 +41,13 @@ from .general import HelpCommand, QuitCommand
 from .open_file import OpenFileCommand
 from .open_gemini_uri import OpenGeminiURICommand
 from .open_other_uri import OpenOtherURICommand
+from .palette_commands import PaletteCommand
 
 ##############################################################################
 COMMANDS: Final[tuple[type[InputCommand], ...]] = (
     HelpCommand,
     QuitCommand,
+    PaletteCommand,
     OpenFileCommand,
     OpenGeminiURICommand,
     OpenOtherURICommand,
@@ -106,15 +108,17 @@ class CommandLine(Vertical):
     a filename or a URL will open that file for viewing; entering a
     directory will open a file opening dialog starting at that location.
 
-    | Command | Aliases | Arguments | Description |
-    | --      | --      | --        | --          |
+    | Command | Aliases | Description |
+    | --      | --      | --          |
     {cli_commands}
 
     ### Special keys
 
     Special keys while in the command line:
     """.format(
-        cli_commands="\n    ".join(sorted(command.help_text() for command in COMMANDS)),
+        cli_commands="\n    ".join(
+            sorted(chain(*(command.help_text() for command in COMMANDS)))
+        ),
     )
 
     BINDINGS = [